refactor: use project-base-directory crate
Instead of using git to determine the base directory of the project, the
logic is outsourced to `project-base-directory`. This checks for the
`PRJ_ROOT` environment variable and will use that if present. Otherwise,
it falls back to using the git repository as the base. However, this
implementation uses gix and so we no longer depend on libgit for doing
so.